requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
flatbuffers==24.3.25
idna==3.10
numpy==1.24.4
openmeteo-requests==1.3.0
openmeteo-sdk==1.18.0
pandas==2.0.3
platformdirs==4.3.6
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
